+++ src/main/ruby/truffleruby/core/truffle/regexp_operations.rb
@@ -18,15 +18,7 @@ module Truffle
       Primitive.regexp_last_match_set(s, v)
     }
 
-    def self.search_check_args(re, str)
-      raise TypeError, 'uninitialized regexp' unless Primitive.regexp_initialized?(re)
-      raise ArgumentError, "invalid byte sequence in #{str.encoding}" unless str.valid_encoding?
-      Primitive.encoding_ensure_compatible(re, str)
-    end
-
     def self.search_region(re, str, start_index, end_index, forward, create_match_data)
-      search_check_args(re, str)
-
       if forward
         from = start_index
         to = end_index
@@ -37,13 +29,7 @@ module Truffle
       match_in_region(re, str, from, to, false, 0, create_match_data)
     end
 
-    def self.search_string(re, str)
-      search_check_args(re, str)
-
-      end_index = str.bytesize
-      match_in_region(re, str, 0, end_index, false, 0)
-    end
-
+    # MRI: rb_reg_match_m/reg_match_pos
     def self.match(re, str, pos=0)
       return nil unless str
 
@@ -55,6 +41,7 @@ module Truffle
       search_region(re, str, pos, str.bytesize, true, true)
     end
 
+    # MRI: rb_reg_match_p
     def self.match?(re, str, pos=0)
       return false unless str